+++ packages/core/src/lib/utils/openapi/default-hook.ts
@@ -2,7 +2,7 @@ import type { Hook } from '@hono/zod-openapi'
 
 import { UNPROCESSABLE_ENTITY } from '../http-status'
 
-const defaultHook: Hook<unknown, unknown, unknown, unknown> = (result, c) => {
+const defaultHook: Hook<any, any, any, any> = (result, c) => {
   if (!result.success) {
     return c.json(
       {